requirements.txt
fastapi
uvicorn
chromadb
google-generativeai
PyPDF2
python-docx
sentence-transformers
requests
fasttext
pydantic
